send position to backend
